+++ projects/np-ui-lib/src/lib/np-data-grid/models/column.model.ts
@@ -18,8 +18,6 @@ export class Column {
         this.onCellClick = item.onCellClick;
         this.styleClass = item.styleClass;
         this.rightAlignText = item.rightAlignText;
-        this.stickyColumnLeft = item.stickyColumnLeft;
-        this.stickyColumnRight = item.stickyColumnRight;
         this.trueFilterText = item.trueFilterText;
         this.falseFilterText = item.falseFilterText;
     }
@@ -38,8 +36,9 @@ export class Column {
     public onCellClick: any;
     public styleClass: string;
     public rightAlignText: boolean;
-    public stickyColumnLeft: boolean;
-    public stickyColumnRight: boolean;
     public trueFilterText: string;
     public falseFilterText: string;
+
+    public _leftPad: number;
+    public _rightPad: number;
 }