+++ projects/np-ui-lib/src/lib/np-tabs/np-tabs.module.ts
@@ -4,10 +4,11 @@ import { NpTabsComponent } from './np-tabs.component';
 import { NpTabComponent } from './np-tab.component';
 import { NpTabContent } from './np-tab-content.directive';
 import { PortalModule } from '@angular/cdk/portal';
+import { ScrollingModule } from '@angular/cdk/scrolling';
 
 @NgModule({
   declarations: [NpTabsComponent, NpTabComponent, NpTabContent],
-  imports: [CommonModule, PortalModule],
+  imports: [CommonModule, PortalModule, ScrollingModule],
   exports: [NpTabsComponent, NpTabComponent, NpTabContent]
 })
 export class NpTabsModule { }