+++ projects/np-ui-lib/CHANGELOG.md
@@ -6,6 +6,7 @@
 # 11.1.3
 - Add npSidepanelContent directive for lazy loading
 - Vertical steps layout update
+- Remove stickyColumnLeft, stickyColumnRight from data grid
 
 # 11.1.2
 - Resolved css issue for vertical line in vertical steps